requirements.txt
acres==0.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.1
distro==1.9.0
docx2txt==0.8
etelemetry==0.3.1
fastapi==0.95.0
fastapi-limiter==0.1.6
filelock==3.17.0
fitz==0.0.1.dev2
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
isodate==0.6.1
jiter==0.8.2
looseversion==1.3.0
lxml==5.3.1
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
openai==0.27.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
prov==2.0.1
puremagic==1.28
pycparser==2.22
pydantic==1.10.7
pydantic_core==2.27.2
pydot==3.0.4
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.10
pytest==7.2.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
redis==5.2.1
requests==2.32.3
scipy==1.15.2
sentencepiece==0.2.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.26.1
tenacity==8.5.0
tokenizers==0.21.0
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.22.0
Wand==0.6.13
yarl==1.18.3
